Renamed canvases to conferences where needed
