+++ server/__main__.py
@@ -3,7 +3,7 @@ import asyncio
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from server.domains import users, canvas
+from server.domains import users, conferences
 from server.database.db_settings import Base, engine
 from server.config import ALLOWED_ORIGINS
 
@@ -12,7 +12,7 @@ Base.metadata.create_all(bind=engine)
 app = FastAPI()
 
 app.include_router(users.router)
-app.include_router(canvas.router)
+app.include_router(conferences.router)
 
 
 app.add_middleware(
